chore(licence): MIT license made more explicit
Based on on the asciidoctor/asciidoctor-backends repo practice.
No explicit individual contributers were highlighted (yet), As I
can only base this on the git commit history, which might be
errornous. So if individual contributers are to be attributed, this
should be done at a later time, by somebody more involved in this
project.