Add npm init step to npm installation
Add instructions to the npm installation section. Create a new folder for slides and then do npm init in that folder. If slides are stored away from the directory where 'node_modules' exists, then there will be an error when attempting to generate slides for a file not found issue.